+++ src/main/java/frc/robot/commands/AutoCommands/AutoLowGoal.java
@@ -2,11 +2,14 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.AutoCommands;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Constants;
+import frc.robot.commands.IndexerCommand;
+import frc.robot.commands.MoveCommand;
+import frc.robot.commands.ShootLow;
 import frc.robot.subsystems.DrivetrainSubsystem;
 import frc.robot.subsystems.ShooterSubsystem;
 import pabeles.concurrency.ConcurrencyOps.NewInstance; 
@@ -18,20 +21,21 @@ import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class AutoHighIntake extends SequentialCommandGroup {
+public class AutoLowGoal extends SequentialCommandGroup {
   /** Creates a new OttoCommand. */
-  public AutoHighIntake(DrivetrainSubsystem drive, ShooterSubsystem shoot, IndexerSubsystem Index, IntakeSubsystem intake) {
+  public AutoLowGoal(DrivetrainSubsystem drive, ShooterSubsystem shoot, IndexerSubsystem Index, IntakeSubsystem intake) {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand()); 
     addCommands(
       new ActuatorDown(intake),
-      new MoveCommand(drive, -3, false),
       new InstantCommand(()-> System.out.println("YOU ARE RUNNING IN AUTO LOL")),
-      //new MoveCommand(drive, 4, true),
-      parallel(new PickupCommmand(intake)), 
+      //comment out above for high goal
       new MoveCommand(drive, 3, true),
-      parallel( new ShootHigh(shoot), new IndexerCommand(Index)),
-      new MoveCommand(drive, -7, false)
+    parallel( new ShootLow(shoot), new IndexerCommand(Index)),
+      new MoveCommand(drive, -8, false)
+      // change above to -8 for low goal
+      // change above to -4 for high hoal
+      //new PickupCommmand(intake)
     );
   }
 }
