+++ src/main/java/frc/robot/commands/AutoCommands/AutoPush.java
@@ -2,25 +2,25 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.AutoCommands;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import edu.wpi.first.wpilibj2.command.WaitCommand;
+import frc.robot.commands.MoveCommand;
+import frc.robot.subsystems.DrivetrainSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
-import edu.wpi.first.wpilibj2.command.InstantCommand;
-import frc.robot.Constants;
+
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class PushAway extends SequentialCommandGroup {
-  /** Creates a new Shoot. */
-  public PushAway(IntakeSubsystem Intake) {
+public class AutoPush extends SequentialCommandGroup {
+  /** Creates a new AutoMove. */
+  public AutoPush(DrivetrainSubsystem drive, IntakeSubsystem intake) {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
-    addCommands( 
-      new InstantCommand( () -> Intake.spinMotor()),
-      new WaitCommand(Constants.fourSeconds),
-      new InstantCommand( () -> Intake.stopMotor())
-      ); 
+    addCommands(
+      new ActuatorDown(intake),
+      parallel(new MoveCommand(drive, -7, false),
+      new AutoPushAway(intake))
+    );
   }
 }
