+++ src/main/java/frc/robot/commands/AutoCommands/DemoAutoRoutine.java
@@ -2,9 +2,13 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.AutoCommands;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import frc.robot.commands.IndexerCommand;
+import frc.robot.commands.MoveCommand;
+import frc.robot.commands.ShootHigh;
+import frc.robot.commands.TurnCommand;
 import frc.robot.subsystems.DrivetrainSubsystem;
 import frc.robot.subsystems.IndexerSubsystem;
 import frc.robot.subsystems.ShooterSubsystem;
@@ -12,9 +16,9 @@ import frc.robot.subsystems.ShooterSubsystem;
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class Spin extends SequentialCommandGroup {
+public class DemoAutoRoutine extends SequentialCommandGroup {
   /** Creates a new AutoMove. */
-  public Spin(DrivetrainSubsystem drive, ShooterSubsystem shoot, IndexerSubsystem Index) {
+  public DemoAutoRoutine(DrivetrainSubsystem drive, ShooterSubsystem shoot, IndexerSubsystem Index) {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
     addCommands(
@@ -23,8 +27,8 @@ public class Spin extends SequentialCommandGroup {
       new TurnCommand(drive, 90),
       new MoveCommand(drive, 8, true),
      new TurnCommand(drive, 90),
-      new MoveCommand(drive, 2, true)
-      //parallel( new ShootHigh(shoot), new IndexerCommand(Index))
+      new MoveCommand(drive, 2, true),
+      parallel( new ShootHigh(shoot), new IndexerCommand(Index))
  
     );
   }
